requirements.txt
PyMuPDF
opencv-python-headless
numpy
Pillow
gradio